+++ saint/HarmoPitchGetter/HarmoPitchHelper.h
@@ -6,7 +6,7 @@
 
 namespace saint {
 bool setIntervalIndex(const std::vector<int> &intervals, size_t *currentIndex,
-                      int tick);
+                      double tick);
 
 std::vector<HarmoNoteSpan>
 toHarmoNoteSpans(const std::vector<MidiNoteMsg> &playedMidiTrack,
